requirements.txt
pytest
selenium
pytest-bdd
allure-pytest